test: replaced avg fuzzer with status fuzzer
